Fix broken side panel in docs
