--- apps/web/src/app/(docs)/docs/layout.tsx
@@ -1,32 +1,0 @@
-import glob from 'fast-glob'
-
-import { Layout } from '@/components/Layout'
-
-import { Section } from '@/components/SectionProvider'
-
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  namespace JSX {
-    interface IntrinsicElements {
-      'chatlio-widget': any;
-    }
-  }
-}
-
-
-export default async function DocsLayout({ children }) {
-  const pages = await glob('**/*.mdx', { cwd: 'src/app/(docs)/docs' })
-  const allSectionsEntries = (await Promise.all(
-    pages.map(async filename => [
-      '/docs/' + filename.replace(/\(docs\)\/?|(^|\/)page\.mdx$/, ''),
-      (await import(`./${filename}`)).sections,
-    ]),
-  )) as Array<[string, Array<Section>]>
-  const allSections = Object.fromEntries(allSectionsEntries)
-
-  return (
-      <Layout allSections={allSections}>
-        {children}
-      </Layout>
-  )
-}
